+++ sdks/opik_optimizer/src/opik_optimizer/utils/__init__.py
@@ -1,0 +1,10 @@
+"""Utility helpers exposed as part of the opik_optimizer package."""
+
+from .core import *  # noqa: F401,F403
+from .dataset_utils import *  # noqa: F401,F403
+from .mcp import *  # noqa: F401,F403
+from .mcp_second_pass import *  # noqa: F401,F403
+from .mcp_simulator import *  # noqa: F401,F403
+from .prompt_segments import *  # noqa: F401,F403
+
+__all__ = []
